+++ lib/logger/error_handler.ex
@@ -1,10 +1,10 @@
 defmodule Logger.ErrorHandler do
   use GenEvent
 
-  # TODO: Make watermark configurable
+  # TODO: Make threshold configurable
   def init({otp?, sasl?}) do
-    {:ok, %{otp: otp?, sasl: sasl?, watermark: 50,
-            mps: 0, last_time: :os.timestamp}}
+    {:ok, %{otp: otp?, sasl: sasl?, threshold: 50,
+            messages: 0, last_time: :os.timestamp, dropped: 0}}
   end
 
   ## Handle event
@@ -39,23 +39,42 @@ defmodule Logger.ErrorHandler do
     do: state
 
   defp log_event(level, kind, data, state) do
-    state = check_watermark(state)
+    state = check_threshold(state)
     Logger.log(level, fn -> format_event(level, kind, data) end)
     state
   end
 
-  defp check_watermark(%{mps: mps, watermark: watermark} = state) when mps <= watermark do
-    %{state | mps: mps + 1}
+  defp check_threshold(%{messages: messages, threshold: threshold} = state)
+      when messages <= threshold do
+    %{state | messages: messages + 1}
   end
 
-  defp check_watermark(%{last_time: last_time} = state) do
+  defp check_threshold(%{last_time: last_time, dropped: dropped} = state) do
     {m, s, _} = now = :os.timestamp
     case last_time do
       {^m, ^s, _} ->
-        # TODO: Drop messages
-        state
+        count = drop_messages(now, 0)
+        %{state | dropped: dropped + count}
       {_, _, _} ->
-        %{state | last_time: now, mps: 0}
+        if dropped > 0 do
+          Logger.warn "Logger dropped #{dropped} OTP/SASL messages that " <>
+                      "exceeded the amount of #{state.threshold} messages/second"
+        end
+        %{state | last_time: now, messages: 1, dropped: 0}
+    end
+  end
+
+  defp drop_messages(now, count) do
+    {m, s, _} = :os.timestamp
+    case now do
+      {^m, ^s, _} ->
+        receive do
+          {:notify, _event} -> drop_messages(now, count + 1)
+        after
+          0 -> count
+        end
+      _ ->
+        count
     end
   end